requirements.txt
# Flask web framework
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3

# WSGI server
gunicorn==21.2.0

# Database
mysql-connector-python==8.1.0

# PDF generation
fpdf2==2.7.6

# Image processing
Pillow==10.3.0
imagehash==4.3.1

# Data processing - compatible versions
numpy==1.24.3
pandas==2.0.3
openpyxl==3.1.2

# Configuration
python-dotenv==1.0.0
